pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pink-agent"
version = "1.0.0"
description = "Cross-platform Telegram agent for remote Claude Code control via voice and text"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "pinkhairedboy", email = "p.nkhaired@gmail.com"}
]
keywords = ["telegram", "claude-code", "voice-control", "remote-control", "ai-assistant"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
]

dependencies = [
    "python-telegram-bot==22.5",
    "python-dotenv==1.0.0",
    "telegramify-markdown==0.5.2",
    "watchfiles>=0.24.0",
    "requests>=2.32.0",
    "psutil>=6.1.0",
    "setproctitle>=1.3.0",
]

[project.urls]
Homepage = "https://github.com/pinkhairedboy/pink-agent"
Repository = "https://github.com/pinkhairedboy/pink-agent"

[project.scripts]
pink-agent = "pink_agent.daemon.supervisor:main"
pink-claude-agent = "pink_agent.claude.agent:main"
pink-telegram-agent = "pink_agent.telegram.agent:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
